changed dtypes for psi_n

+++ src/gaussian_schell.py
@@ -54,11 +54,11 @@ def psi_n(
 
     const_factor = ((2 * c) / torch.pi)**0.25
     
-    log_n_factor = -0.5 * (n_col * torch.log(torch.tensor(2, dtype=torch.float32, device=narr.device)) + torch.tensor(gammaln(n_col.cpu() + 1), dtype=torch.complex64, device=narr.device))
+    log_n_factor = -0.5 * (n_col * torch.log(torch.tensor(2, dtype=torch.float64, device=narr.device)) + torch.tensor(gammaln(n_col.cpu() + 1), dtype=torch.complex128, device=narr.device))
     n_factor = torch.exp(log_n_factor)
 
-    x_scaled = x * torch.sqrt(torch.tensor(2 * c, dtype=torch.float32, device=narr.device))
-    hermite_term = torch.tensor(eval_hermite(n_col.cpu(), x_scaled.cpu()), dtype=torch.complex64, device=narr.device)
+    x_scaled = x * torch.sqrt(torch.tensor(2 * c, dtype=torch.float64, device=narr.device))
+    hermite_term = torch.tensor(eval_hermite(n_col.cpu(), x_scaled.cpu()), dtype=torch.complex128, device=narr.device)
 
     exp_term = torch.exp(-c * x**2)
